Configure docker to have services-dev and -test
With these settings, we should be able to run our last release on
services-test while the latest commit runs on services-dev. The builder
will need changes in another PR to work as intended.

+++ deployment/update.sh
@@ -1,4 +1,15 @@
 #!/usr/bin/env bash
-docker-compose build
+set -e
+case `hostname` in
+	services-stable)
+		COMPOSE_FILE="docker-compose.yml"
+		docker pull raidennetwork/raiden-services:stable
+		;;
+	services-dev)
+		COMPOSE_FILE="docker-compose.yml -f docker-compose.latest.yml"
+		docker pull raidennetwork/raiden-services:latest
+		;;
+esac
+[ -z $COMPOSE_FILE ] && { echo "Running on unknown host"; exit 1; }
 docker-compose down
-docker-compose -f docker-compose.yml up -d
+docker-compose -f $COMPOSE_FILE up -d
